requirements.txt
anyio==3.5.0
asgiref==3.4.1
asn1crypto==0.24.0
autopep8==1.6.0
blinker==1.4
certifi==2018.8.24
chardet==3.0.4
click==8.0.3
cloud-init==20.4.1
configobj==5.0.6
cryptography==2.6.1
distro-info==0.21
elara==0.5.4
entrypoints==0.3
fastapi==0.71.0
h11==0.12.0
httptools==0.3.0
idna==3.3
importlib-metadata==4.10.0
Jinja2==2.10
jsonpatch==1.21
jsonpointer==1.10
jsonschema==2.6.0
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.0
msgpack==1.0.3
oauthlib==2.1.0
pycodestyle==2.8.0
pycrypto==2.6.1
pycurl==7.43.0.2
pydantic==1.9.0
PyGObject==3.30.4
PyJWT==1.7.0
python-apt==1.8.4.3
python-dotenv==0.19.2
pyxdg==0.25
PyYAML==6.0
requests==2.21.0
safer==4.4.0
schedule==1.1.0
SecretStorage==2.3.1
six==1.12.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing-extensions==4.0.1
unattended-upgrades==0.1
urllib3==1.24.1
uvicorn==0.17.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
zipp==3.7.0
